requirements.txt
groq==0.9.0
httpx==0.27.0
pydantic==2.8.2
Pillow==10.4.0
python-dotenv==1.0.1
fastmcp==0.1.0
readabilipy==0.2.0
markdownify==0.11.6
beautifulsoup4==4.12.3